requirements.txt
aiogram==3.22.0
aiogram-dialog==2.4.0
PyYAML==6.0.2
redis==6.4.0
openai==1.100.2
fluentogram==1.2.1
watchdog==6.0.0
git+https://github.com/molchanov-vs/my_tools.git@bb255b7